Fix circular dependency and refactor Log
